requirements.txt
streamlit
google-generativeai
requests
youtube-transcript-api
yt-dlp
python-dotenv